@@ -23,6 +23,84 @@ describe('regex-tagger should work', function () {
     cy.get('mat-option').contains('integration_test_project').click();
   }
 
+
+  function tagDoc() {
+    cy.get('.cdk-column-actions:nth(1)').should('be.visible').click('left');
+    cy.get('[data-cy=appRegexTaggerMenuTagDoc]').should('be.visible').click();
+    cy.fixture('sample_doc').then(sampleDoc => {
+      let json = JSON.stringify(sampleDoc);
+      cy.get('[data-cy=appRegexTaggerTagDocDialogDocument] input:first()').should('be.visible').click()
+        .clear().invoke('val', json).trigger('change');
+      cy.get('[data-cy=appRegexTaggerTagDocDialogDocument]').click().type(' ');
+    });
+    cy.wait(300);
+    cy.get('[data-cy=appRegexTaggerTagDocDialogFields]').click().then((field => {
+      cy.wrap(field).should('have.class', 'mat-focused');
+      cy.closeCurrentCdkOverlay();
+      cy.matFormFieldShouldHaveError(field, 'required');
+      cy.wrap(field).click();
+      cy.get('.mat-option-text').contains('comment_content').scrollIntoView().should('be.visible').click();
+      cy.closeCurrentCdkOverlay();
+      cy.wrap(field).find('mat-error').should('have.length', 0)
+    }));
+    cy.get('.mat-dialog-container [type="submit"]').should('be.visible').click();
+    cy.wait('@postRegexTaggers').then(resp => {
+      expect(resp.status).to.eq(200);
+      expect(resp.response.body.matches.length).to.eq(1, 'should have found a match');
+    });
+    cy.get('.code-wrapper').should('be.visible');
+    cy.get('[data-cy=appRegexTaggerTagDocDialogClose]').click();
+  }
+
+  function tagRandomDoc() {
+    cy.get('.cdk-column-actions:nth(1)').should('be.visible').click('left');
+    cy.get('[data-cy=appRegexTaggerMenuTagRandomDoc]').should('be.visible').click();
+    cy.get('[data-cy=appRegexTaggerTagRandomDocDialogIndices]').click().then((indices => {
+      cy.wrap(indices).should('have.class', 'mat-focused');
+      cy.closeCurrentCdkOverlay();
+      cy.matFormFieldShouldHaveError(indices, 'required');
+      cy.wrap(indices).click();
+      cy.get('.mat-option-text:nth(0)').should('be.visible').click();
+      cy.closeCurrentCdkOverlay();
+      cy.wrap(indices).find('mat-error').should('have.length', 0)
+    }));
+    cy.get('[data-cy=appRegexTaggerTagRandomDocDialogfields]').click().then((fields => {
+      cy.wrap(fields).should('have.class', 'mat-focused');
+      cy.closeCurrentCdkOverlay();
+      cy.matFormFieldShouldHaveError(fields, 'required');
+      cy.wrap(fields).click();
+      cy.get('.mat-option-text:nth(1)').should('be.visible').click();
+      cy.closeCurrentCdkOverlay();
+      cy.wrap(fields).find('mat-error').should('have.length', 0)
+    }));
+
+    cy.get('.mat-dialog-container [type="submit"]').should('be.visible').click();
+    cy.wait('@postRegexTaggers').then(resp => {
+      expect(resp.status).to.eq(200);
+    });
+    cy.get('.code-wrapper').should('be.visible');
+    cy.get('[data-cy=appRegexTaggerTagRandomDocDialogClose]').click();
+
+  }
+
+
+  function tagText() {
+    cy.get('.cdk-column-actions:nth(1)').should('be.visible').click('left');
+    cy.get('[data-cy=appRegexTaggerMenuTagText]').should('be.visible').click();
+    cy.get('[data-cy=appRegexTaggerTagTextDialogText] input:first()').should('be.visible').click()
+      .clear().invoke('val', 'tere Ooot, misasi see võõrast aiast maasikate võtmine on, kui ta vargus pole???').trigger('change');
+    cy.get('[data-cy=appRegexTaggerTagTextDialogText]').click().type(' ');
+
+    cy.get('.mat-dialog-container [type="submit"]').should('be.visible').click();
+
+    cy.wait('@postRegexTaggers').then(resp => {
+      expect(resp.status).to.eq(200);
+      expect(resp.response.body.length).to.eq(1, 'should have found a match');
+    });
+    cy.get('.code-wrapper').should('be.visible');
+    cy.get('[data-cy=appRegexTaggerTagTextDialogClose]').click();
+  }
+
   it('should be able to create a new regex-tagger', function () {
     // create clustering
     initPage();
@@ -35,7 +113,7 @@ describe('regex-tagger should work', function () {
     cy.get('[data-cy=appRegexTaggerCreateDialogLexicon]').click().then((desc => {
       cy.wrap(desc).should('have.class', 'mat-focused').type('b').find('textarea').clear();
       cy.matFormFieldShouldHaveError(desc, 'required');
-      cy.wrap(desc).type('test');
+      cy.wrap(desc).type('tere');
     }));
 
     cy.get('[data-cy=appRegexTaggerCreateDialogSubmit]').click();
@@ -55,5 +133,17 @@ describe('regex-tagger should work', function () {
       expect(tag.status).to.eq(200);
     });
     cy.get('.code-wrapper').should('be.visible');
+    cy.get('[data-cy=appRegexTaggerMultiTagDialogClose]').click();
+
+
+    tagDoc();
+    tagRandomDoc();
+    tagText();
+
+    cy.get('.cdk-column-actions:nth(1)').click('left');
+    cy.get('[data-cy=appRegexTaggerMenuDelete]').click();
+    cy.get('.mat-dialog-container [type="submit"]').should('be.visible').click();
+    cy.get('.cdk-column-actions').should('have.length', 1);
+
   });
 });
