@@ -17,6 +17,9 @@ import {CreateRegexTaggerDialogComponent} from './create-regex-tagger-dialog/cre
 import {MultiTagTextDialogComponent} from './multi-tag-text-dialog/multi-tag-text-dialog.component';
 import {HttpErrorResponse} from '@angular/common/http';
 import {EditRegexTaggerDialogComponent} from './edit-regex-tagger-dialog/edit-regex-tagger-dialog.component';
+import {TagTextDialogComponent} from './tag-text-dialog/tag-text-dialog.component';
+import {TagDocDialogComponent} from './tag-doc-dialog/tag-doc-dialog.component';
+import {TagRandomDocComponent} from './tag-random-doc/tag-random-doc.component';
 
 @Component({
   selector: 'app-regex-tagger',
@@ -28,14 +31,13 @@ import {EditRegexTaggerDialogComponent} from './edit-regex-tagger-dialog/edit-re
 })
 export class RegexTaggerComponent implements OnInit, OnDestroy, AfterViewInit {
   public displayedColumns = ['select', 'id', 'description', 'operator', 'matchType', 'requiredWords', 'phraseSlop',
-    'counterSlop', 'nAllowedEdits', 'fuzzy', 'ignoreCase', 'ignorePunctuation', 'edit'];
+    'counterSlop', 'nAllowedEdits', 'fuzzy', 'ignoreCase', 'ignorePunctuation', 'actions'];
   public isLoadingResults = true;
   public tableData: MatTableDataSource<RegexTagger> = new MatTableDataSource();
 
   destroyed$ = new Subject<boolean>();
   selectedRows = new SelectionModel<RegexTagger>(true, []);
   filteredSubject = new Subject();
-  resultsLength: number;
   expandedElement: RegexTagger | null;
   currentProject: Project;
   patchRowQueue: Subject<RegexTagger> = new Subject();
@@ -50,9 +52,6 @@ export class RegexTaggerComponent implements OnInit, OnDestroy, AfterViewInit {
   }
 
   ngOnInit(): void {
-    this.tableData.sort = this.sort;
-    this.tableData.paginator = this.paginator;
-
     this.patchRowQueue.pipe(takeUntil(this.destroyed$), debounceTime(50)).subscribe(row => {
       if (this.currentProject) {
         this.regexTaggerService.patchRegexTagger(this.currentProject.id, row.id, row).subscribe();
@@ -62,6 +61,8 @@ export class RegexTaggerComponent implements OnInit, OnDestroy, AfterViewInit {
 
 
   ngAfterViewInit(): void {   // If the user changes the sort order, reset back to the first page.
+    this.tableData.sort = this.sort;
+    this.tableData.paginator = this.paginator;
     this.sort.sortChange.subscribe(() => this.paginator.pageIndex = 0);
 
     merge(this.sort.sortChange, this.paginator.page, this.filteredSubject)
@@ -82,7 +83,6 @@ export class RegexTaggerComponent implements OnInit, OnDestroy, AfterViewInit {
         this.logService.snackBarError(data, 2000);
       } else if (data) {
         this.isLoadingResults = false;
-        this.resultsLength = data.count;
         this.tableData.data = data.results;
       }
     });
@@ -129,7 +129,33 @@ export class RegexTaggerComponent implements OnInit, OnDestroy, AfterViewInit {
       }
     });
   }
+  onTagText(element: RegexTagger): void {
+    this.dialog.open(TagTextDialogComponent, {
+      maxHeight: '750px',
+      width: '700px',
+      disableClose: true,
+      data: {currentProjectId: this.currentProject.id, tagger: element}
+    });
+  }
 
+
+  onTagDoc(element: RegexTagger): void {
+    this.dialog.open(TagDocDialogComponent, {
+      maxHeight: '750px',
+      width: '700px',
+      disableClose: true,
+      data: {currentProjectId: this.currentProject.id, tagger: element}
+    });
+  }
+
+  onTagRandomDoc(element: RegexTagger): void {
+    this.dialog.open(TagRandomDocComponent, {
+      maxHeight: '750px',
+      width: '700px',
+      disableClose: true,
+      data: {currentProjectId: this.currentProject.id, tagger: element}
+    });
+  }
   openMultiTagDialog(): void {
     this.dialog.open(MultiTagTextDialogComponent, {
       maxHeight: '90vh',
@@ -176,6 +202,24 @@ export class RegexTaggerComponent implements OnInit, OnDestroy, AfterViewInit {
         }
       });
     }
+  }
+
+  onDelete(cluster: RegexTagger, index: number): void {
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: {confirmText: 'Delete', mainText: 'Are you sure you want to delete this regex tagger?'}
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+
+        const body = {ids: [index]};
+        this.regexTaggerService.bulkDeleteRegexTaggers(this.currentProject.id, body).subscribe(() => {
+          this.logService.snackBarMessage(`Deleted regex tagger ${cluster.description}`, 2000);
+          this.tableData.data.splice(index, 1);
+          this.tableData.data = [...this.tableData.data];
+        });
+      }
+    });
   }
 
   removeSelectedRows(): void {
